Extract `tmt run` from `tmt.cli` into standalone module
